+++ lib/hexagon.dart
@@ -1,0 +1,7 @@
+library hexagon;
+
+export 'src/grid/hexagon_offset_grid.dart';
+export 'src/hexagon_clipper.dart';
+export 'src/hexagon_painter.dart';
+export 'src/hexagon_type.dart';
+export 'src/hexagon_widget.dart';